fix(#1614): clear 9 ban-scanner violations breaking HEAD build
- flex_jet.rs: bundle 19-arg hand_vgh/jet_vgh oracles into a Row struct
  (drops 2x #[allow(clippy::too_many_arguments)]); test still asserts <=1e-12.
- delete 3 #[ignore] timing microbenches (gradient_paths, cell_moment_assembly,
  row_jet_program) that asserted nothing and were env-dependent; the
  correctness oracles beside them stay. Drops the #[test]-without-assertions
  violation too. Remove now-dead softmax_fixture_k helper.
- multinomial_reml.rs: move 2 #[cfg(test)] impl methods into the test module's
  own impl MultinomialFamily block.
- split 4 logdet-adjoint tests out of manifold/tests.rs (10012>10000 lines)
  into sibling tests_logdet_adjoint_780.rs, declared in mod.rs.
