+++ crates/gam-sae/src/row_jet_program.rs
@@ -1389,194 +1389,6 @@ mod tests {
         (prog, inv_tau)
     }
 
-    /// Parametrized softmax fixture with `n_atoms` softmax atoms, each carrying a
-    /// free logit primary and `latent_dim` free coord primaries, so
-    /// `n_primaries = n_atoms·(1 + latent_dim)`. Layout: logit slots
-    /// `0..n_atoms`, then atom `k`'s coord axis `j` at `n_atoms + k·latent_dim +
-    /// j`. Used by the #932 ns/row microbench to instantiate the tower at
-    /// `K = n_primaries` for `K ∈ {8, 16}` (the softmax gate Hessian is `n_atoms³`,
-    /// the cost driver the hand path pays per output column).
-    fn softmax_fixture_k(
-        n_atoms: usize,
-        latent_dim: usize,
-        n_basis: usize,
-        out_dim: usize,
-        inv_tau: f64,
-    ) -> SaeReconstructionRowProgram {
-        let mk_atom = |seed: f64| {
-            let phi: Vec<f64> = (0..n_basis)
-                .map(|b| 0.3 + 0.2 * (b as f64 + seed))
-                .collect();
-            let d_phi: Vec<Vec<f64>> = (0..n_basis)
-                .map(|b| {
-                    (0..latent_dim)
-                        .map(|axis| 0.1 * (b as f64 + 1.0) - 0.05 * axis as f64 + 0.03 * seed)
-                        .collect()
-                })
-                .collect();
-            let d2_phi: Vec<Vec<Vec<f64>>> = (0..n_basis)
-                .map(|b| {
-                    (0..latent_dim)
-                        .map(|a| {
-                            (0..latent_dim)
-                                .map(|bb| {
-                                    0.02 * (b as f64 + 1.0)
-                                        + 0.01 * (a as f64)
-                                        + 0.01 * (bb as f64)
-                                        + 0.004 * seed
-                                })
-                                .collect()
-                        })
-                        .collect()
-                })
-                .collect();
-            let decoder: Vec<Vec<f64>> = (0..n_basis)
-                .map(|b| {
-                    (0..out_dim)
-                        .map(|c| 0.5 - 0.1 * (b as f64) + 0.07 * (c as f64) + 0.02 * seed)
-                        .collect()
-                })
-                .collect();
-            AtomRowBasisJet {
-                phi,
-                d_phi,
-                d2_phi,
-                decoder,
-                latent_dim,
-            }
-        };
-        let logits: Vec<f64> = (0..n_atoms)
-            .map(|k| 0.4 - 0.13 * k as f64 + 0.05 * (k as f64).sin())
-            .collect();
-        let e: Vec<f64> = logits.iter().map(|&l| (l * inv_tau).exp()).collect();
-        let s: f64 = e.iter().sum();
-        let gate_value: Vec<f64> = e.iter().map(|&v| v / s).collect();
-        let atoms: Vec<AtomRowBasisJet> = (0..n_atoms).map(|k| mk_atom(k as f64)).collect();
-        let logit_slot: Vec<Option<usize>> = (0..n_atoms).map(Some).collect();
-        let coord_slot: Vec<Vec<usize>> = (0..n_atoms)
-            .map(|k| (0..latent_dim).map(|j| n_atoms + k * latent_dim + j).collect())
-            .collect();
-        SaeReconstructionRowProgram {
-            atoms,
-            gate_value,
-            logits,
-            gate_scale: vec![1.0; n_atoms],
-            gate_shift: vec![0.0; n_atoms],
-            gate: RowGate::Softmax { inv_tau },
-            logit_slot,
-            coord_slot,
-            n_primaries: n_atoms * (1 + latent_dim),
-        }
-    }
-
-    /// #932 ns/row microbench: the production packed jet recon
-    /// ([`SaeReconstructionRowProgram::reconstruction_all_columns_packed`], gate +
-    /// basis jets HOISTED out of the column loop, softmax denom/recip SHARED)
-    /// versus the hand path ([`hand_softmax_column`], the old `row_jets_for_logdet`
-    /// closed-form softmax gate Jacobian/Hessian × decoded basis, re-derived per
-    /// output column). Both produce every output column's value/grad/Hessian. This
-    /// is a measurement, not an assertion; run with `--release --nocapture`.
-    #[test]
-    #[ignore = "microbench; run with --release --nocapture"]
-    fn bench_recon_jet_vs_hand_ns_per_row() {
-        let out_dim = 16;
-        let n_basis = 4;
-        let inv_tau = 1.3;
-        let warm = 2_000usize;
-        let iters = 50_000usize;
-
-        // K=8: 4 atoms × (1 logit + 1 coord) = 8 primaries.
-        bench_one::<8>(
-            softmax_fixture_k(4, 1, n_basis, out_dim, inv_tau),
-            inv_tau,
-            warm,
-            iters,
-            "K=8 (4 atoms, latent_dim=1)",
-        );
-        // K=16: 8 atoms × (1 logit + 1 coord) = 16 primaries.
-        bench_one::<16>(
-            softmax_fixture_k(8, 1, n_basis, out_dim, inv_tau),
-            inv_tau,
-            warm,
-            iters / 2,
-            "K=16 (8 atoms, latent_dim=1)",
-        );
-    }
-
-    fn bench_one<const K: usize>(
-        prog: SaeReconstructionRowProgram,
-        inv_tau: f64,
-        warm: usize,
-        iters: usize,
-        label: &str,
-    ) {
-        use std::hint::black_box;
-        use std::time::Instant;
-        let out_dim = prog.out_dim();
-
-        // Correctness gate before timing: the packed jet all-columns path agrees
-        // with the hand path on value/grad/Hessian (the #932 bit-identity bar).
-        let cols = prog.reconstruction_all_columns_packed::<K>();
-        for c in 0..out_dim {
-            let hand = hand_softmax_column(&prog, c, inv_tau);
-            let h_floor = hand
-                .second
-                .iter()
-                .flatten()
-                .fold(0.0_f64, |m, x| m.max(x.abs()));
-            assert!((cols[c].value() - hand.value).abs() <= 1e-9 * hand.value.abs().max(1.0));
-            for a in 0..K {
-                for b in 0..K {
-                    assert!(
-                        (cols[c].h()[a][b] - hand.second[a][b]).abs()
-                            <= 1e-8 * h_floor.max(1e-12)
-                    );
-                }
-            }
-        }
-
-        // Warmup.
-        for _ in 0..warm {
-            black_box(prog.reconstruction_all_columns_packed::<K>());
-            for c in 0..out_dim {
-                black_box(hand_softmax_column(&prog, c, inv_tau));
-            }
-        }
-
-        // Jet: one all-columns packed call per row.
-        let t0 = Instant::now();
-        for _ in 0..iters {
-            black_box(prog.reconstruction_all_columns_packed::<K>());
-        }
-        let jet_ns = t0.elapsed().as_nanos() as f64 / iters as f64;
-
-        // Hand: every output column re-derived (the production hand path).
-        let t1 = Instant::now();
-        for _ in 0..iters {
-            for c in 0..out_dim {
-                black_box(hand_softmax_column(&prog, c, inv_tau));
-            }
-        }
-        let hand_ns = t1.elapsed().as_nanos() as f64 / iters as f64;
-
-        // Jet per-column (no hoist) to isolate the hoist/share win.
-        let t2 = Instant::now();
-        for _ in 0..iters {
-            for c in 0..out_dim {
-                black_box(prog.reconstruction_column_packed::<K>(c));
-            }
-        }
-        let jet_percol_ns = t2.elapsed().as_nanos() as f64 / iters as f64;
-
-        eprintln!(
-            "[#932 bench {label}] out_dim={out_dim}  jet(hoisted)={jet_ns:.1} ns/row  \
-             jet(per-col)={jet_percol_ns:.1} ns/row  hand={hand_ns:.1} ns/row  \
-             jet-vs-hand={:.2}x  hoist={:.2}x",
-            hand_ns / jet_ns,
-            jet_percol_ns / jet_ns
-        );
-    }
-
     /// INDEPENDENT scalar witness for the reconstruction column `ẑ_c(δ)` as a
     /// function of the primary-increment vector `δ` (the displacement of each
     /// tower primary from its seed value: a coord primary seeds at value 0, a
